--- frontend/src/components/Modal2/Modal.scss
@@ -1,37 +1,0 @@
-.modal__container {
-  position: fixed;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  top: 0;
-  left: 0;
-  height: 100vh;
-  width: 100vw;
-  background-color: rgba(0, 0, 0, 0.42);
-}
-
-.modal__content {
-  position: relative;
-  background-color: white;
-  border-radius: 12px;
-  padding: 50px 30px;
-  height: max-content;
-  width: max-content;
-}
-
-.modal__button_close {
- position: absolute;
- right: -50px;
- top: 5px; 
- border: none;
- border-radius: 50%;
- height: 35px;
- width: 35px;
- font-size: 20px;
- background-color: rgba(255, 255, 255, 0.34);
- cursor: pointer;
-}
-
-.modal__button_close:hover {
-  background-color: rgba(255, 255, 255, 0.644);
-}
