+++ frontend/src/components/SignUp/SignUp.scss
@@ -1,18 +1,3 @@
-.signup__button {
-  padding: 0 10px;
-  font-size: 15px;
-  color: white;
-  background: none;
-  border: none;
-  cursor: pointer;
-  &:hover {
-    opacity: 0.8;
-  }
-  &:active {
-    background-color: rgba(210, 210, 210, 0.577);
-  }
-}
-
 .signup__form {
   display: flex;
   flex-direction: column;
